Move animation timings into constants file and use Animated.parallel
