+++ app/constants/timings.js
@@ -1,0 +1,3 @@
+export const profileOpen   = 333
+export const profileClose  = 333
+export const profileSwitch = 333
